Update patch set 1
Patch Set 1:

All BSD systems have relatively low limits on command line length, or at least they used to.  (And, for what it's worth, the total size of environment variables also counts against this limit.)  It seems to me that we may as well just do this on all systems.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>