Update patch set 4
Patch Set 4:

(3 comments)

Thanks for the review!

I'll be happy to have this in 1.15, if possible.

I share your concern about the loss of deduping when there are many unique labels, but felt the benefits outweighed the costs here. I considered adding the labels to a new debug-level 3, but decided that would be confusing since it would be a higher-debug level for an aggregated/less-verbose presentation.

Patch-set: 4