[dev.link] cmd/link/internal/loader: update ext reloc type in Reloc2.SetType
Clients may neede to invoke the loader.Reloc2.SetType method to reset
the type of a relocation from external flavor to internal flavor,
meaning that the external type add-in needs to be zeroed (this is
needed when adding dynsym entries).

Update loader.Reloc2 to store the external reloc by reference instead
of value, so that loader.Reloc2 can update the underlying value.

Change-Id: I8e26cdae0a0f353019acba5f9c8a0506e3970266