Update patch set 1
Patch Set 1:

(2 comments)

Just checking my understanding: it looks like this puts a fingerprint in both the export data and the object header. Is that necessary because in some configurations we split the part of the object for the compiler and the part for the linker, and both need the fingerprint?

Patch-set: 1