Update patch set 5
Patch Set 5:

Thanks for the comments.

I believe in your CL to convert the symtab phase you wrote:

"This is more or less a direct translation, to get things going.
There are more things we can do to make it better..."

This is more or less what I have been doing in converting dodata(). 

If folks would prefer that I slow down and try to do more refactoring and problem fixing along the way (similar to the work I did in dodata earlier to reduce code duplication and improve readability), I can certainly go that route -- it's just a question of what the priorities are.

Patch-set: 5
Reviewer: Gerrit User 14020 <14020@62eb7196-b449-3ce5-99f1-c037f21e1705>