crypto/x509: add CertPool.AddCertFunc
Allows building CertPools that consume less memory. (Most certs are
never accessed. Different users/programs access different ones, but
not many.)

This CL only adds the new mechanism (and uses it internally for the
old AddCert) but does not modify any existing root pool behavior.
(That is, the default Unix roots are still all slurped into memory as
of this CL)

Change-Id: Ib3a42e4050627b5e34413c595d8ced839c7bfa14