Update patch set 2
Patch Set 2:

This needs a proposal.

I am not sure the complexity is worth it, or worth exposing. For the system roots, we can just do this behind the scenes. For custom roots, it can be done by the application by looking at the chain on the fly.

Patch-set: 2
Reviewer: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>