Update patch set 1
Patch Set 1:

> Patch Set 1: TryBot-Result-1
> 
> 1 of 20 TryBots failed:
> Failed on misc-compile-netbsd: https://storage.googleapis.com/go-build-log/21c1c482/misc-compile-netbsd_0f10b38b.log
> 
> Consult https://build.golang.org/ to see whether they are new failures. Keep in mind that TryBots currently test *exactly* your git commit, without rebasing. If your commit's git parent is old, the failure might've already been fixed.

I've been trying to reproduce this failure this evening as well without success.  What's interesting is that rebasing hasn't really helped.  The same sorts of failures can be seen with the patch tested earlier in the week that was rebased on master at the start of the month, the patch rebased this morning and this patch rebased just now.  Maybe there really is an issue here.

Patch-set: 1
CC: Gerrit User 19185 <19185@62eb7196-b449-3ce5-99f1-c037f21e1705>