Update patch set 2
Patch Set 2:

What will happen when this is merged to master?  We don't want a commit with a binary in master at all, since it grows the git repo for everyone.  In the past we've deliberately trashed the git repo to avoid having binaries in the history; should we do that here?

Patch-set: 2
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>