Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> I want to note that the error returned in this case formats as "png: unsupported feature: dimension overflow", which seems not entirely clear (overflow is a feature??). Not sure if it can be changed now though.

Yeah, the error message isn't great, but it is an UnsupportedError, in that the PNG is valid but this implementation cannot decode it. As for changing the UnsupportedError string prefix, at this point https://www.hyrumslaw.com/ possibly applies.

Patch-set: 2
Reviewer: Gerrit User 5899 <5899@62eb7196-b449-3ce5-99f1-c037f21e1705>