go/internal/srcimporter: disable TestCgo on NetBSD
This test was added in CL 33677 to test newly added functionality
there, but the test fails on NetBSD. The underlying issue exists on
other OSes, and is expected to be fixed shortly, but this CL disables
the test on NetBSD to turn the tree green again, and so the
functionality can remain in tree and testable.

Updates #38649.

Change-Id: I73230a0c3505ec11c9706a31eafa7c6af8352926