hash/crc32: simplify hasVX checking on s390x
Originally, we use an assembly function that returns a boolean result to
tell whether the machine has vector facility or not. It is now no longer
needed when we can directly use cpu.S390X.HasVX variable. This CL
also removes the last occurence of hasVectorFacility function on s390x.

DO NOT SUBMIT(last statement is not true yet until the other two gets merged)
Change-Id: Id20cb746c21eacac5e13344b362e2d87adfe4317