Update patch set 1
Patch Set 1:

Please open an issue to discuss this. Add a line like "Fixes #12345" that links to the issue from the top PR comment.

In this situation, the error message is probably correct. v0.0.0 is not a real version tag. Kubernetes uses those with replace directives, which only apply within the modules they appear in. Those modules (in particular, k8s.io/kubernetes) are not intended to be used outside the Kubernetes ecosystem.

That said, I think this change should go in. Ideally there should be a test in testdata/script. As a starting point:

    [!exec:git] skip
    ! go list -m all
    stderr 'unable to connect'

    -- go.mod --
    module m

    go 1.14

    require 0.0.0.0/foo.git v0.0.0

Patch-set: 1
CC: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>
CC: Gerrit User 10033 <10033@62eb7196-b449-3ce5-99f1-c037f21e1705>
CC: Gerrit User 17092 <17092@62eb7196-b449-3ce5-99f1-c037f21e1705>