Update patch set 6
Patch Set 6:

looking good.

do we have tests that we are in fact evaluating the rhs of such initializations for side-effects?

they’ll be annoying to write, but we should have them. I wonder whether we should add a new verb to the test/run.go tests to say that we expect a panic. the runner can fail if the exit code is 0 or if the output doesn’t include the word “panic”. then you can simply write things that should panic. in fact, I’m not sure you could do the test without that, since some of the tests that we need to write happen at the top level, and there’s no way to catch such panics in the same test. just musing our loud on my phone...

(btw, if anyone knows why/how Gerrit mobile disables autocapitalization on iOS, it’d be nice to fix that!)

Patch-set: 6
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>