runtime: make addrRange[s] operate on linearized addresses
Currently addrRange and addrRanges operate on real addresses (that is,
the addresses they manipulate don't include arenaBaseOffset). While this
is generally OK because even those platforms which have a segmented
address space usually don't give addresses in a different segment, today
it causes a mismatch between the scavenger and the rest of the page
allocator. The scavenger scavenges from the highest addresses first, but
only via real address, whereas the page allocator allocates memory in
linearized address order.

So this change makes addrRange and addrRanges, i.e. what the scavenger
operates on, use linearized addresses. However, lots of the page
allocator relies on an addrRange containing real addresses.

To make this transition less error-prone, this change introduces a new
type, linearAddress, whose purpose is to make linearized addresses
a distinct type, so any attempt to trivially mix real and linearized
addresses will cause a compilation error. It also adds a bunch of
trivial convenience methods to ease the transition to the new type.

This change doesn't attempt to use linearAddress in all of the runtime,
a follow-up change will look for and catch remaining uses of a
linearized address which doesn't use the type.

Updates #35788.

Change-Id: I991d891ac8ace8339ca180daafdf6b261a4d43d1