runtime: use linearAddress in more parts of the runtime
This change uses the new linearAddress type, which represents an address
in a linearized view of the address space for platforms whose address
spsace is segmented, in more parts of the runtime to improve readability
and type safety.

This change also adds some of the linearAddress-related functions to the
"must be inlined" test list so that performance in e.g. arena lookup,
which is on a very hot path in the runtime, doesn't regress.

Updates #35788.

Change-Id: I4e1c5fed3ed68aa12f49a42b82eb3f46aba82fc1