runtime: avoid overflow when representing address ranges
Currently there are a number of places in the runtime where we manage a
range of addresses using an inclusive lower bound and an exclusive upper
bound. Unfortunately this has serious problems when the upper part of
that bound is at the top of the address space, and wraps around. This
case is possible on any platform, but much more likely on 32-bit
platforms.

For cases where we manage a range of real addresses, either use
inclusive upper bounds or base-relative offsets instead. This includes
addrRange, linearAlloc, curArena, and a number of places in pageAlloc.

Fixes #35954.

Change-Id: Ifec497f2fc23f4d96acc8c124942cfff7737ca7d