strconv: implement parseFloat returning no. of bytes consumed (experiment)
DO NOT REVIEW. DO NOT SUBMIT.

This is just an experiment to see how hard it would be to have an
internal version of parseFloat that also returns the number of
bytes consumed. This should make it easier to implement parsing
complex numbers (ParseComplex).

A handful of tests still fail. Haven't investigated yet.

Change-Id: I8ff66b582ae8b468d89c9ffc35c569c735cf0341