Update patch set 26
Patch Set 26:

> Patch Set 26:
> 
> > Patch Set 22: Code-Review-1
> > 
> > (6 comments)
> 
> Can at least FormatComplex be released in Go1.15?

FormatComplex is trivial to write even by hand. It doesn't really make a lot of sense to provide the easy one and not the other one.

We should try to get ParseComplex in, and I think it's possible, but I don't have the time to care of it. We've lived without it for a decade, it will still be ok if it goes one release later. This is not urgent.

Patch-set: 26
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>