Update patch set 8
Patch Set 8:

rsc wrote in #10848: "That map thing is a terrible idiom and should not be encouraged." I don't think anything has changed since then.

If there's real-world code that will benefit from this, that would be useful data to include in this conversation. But as-is, I'm skeptical that the extra compiler code complexity is net positive.

Patch-set: 8