Update patch set 2
Patch Set 2:

For the gscan/hchan special case, I wonder if we could keep the lock rank checking generic and instead add something to syncadjustsudogs to inform the lock rank checker that it's doing something funky. That would reduce the "action at a distance" and keep the knowledge of this funny lock ordering localized to syncadjustsudogs.

For example, syncadjustsudogs could call something in the lock rank code that would find the lockRankGscan it's holding and rewrite that held rank to some fake rank < lockRankHchan. Then the rank checking doesn't have to do anything differently.

Or would it work for syncadjustsudogs to call lockWithRank instead of lock(&sg.c.lock) and just provide a different rank for the hchan acquisition?

(There may be better ways; these are just some that come to mind.)

Patch-set: 2