Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Also had a failure in the overnight run, which also was undertesting because I forgot to 'go clean -cache' between iterations.  Perhaps it was caused by that typo -- will try again with the fix, aimed at the failure case (in runtime): 

This is looking good to me.  I've been testing v2 of the patch locally with the typo fixed for five hours now.  The patch seems to be functioning correctly, in that instructions are being correctly padded, and I haven't seen any crashes, even when testing with GOGC=1 and using taskset 01.  Having said that, I've been unable to reproduce the runtime crash you reported on my machine, even with the version of the patch with the typo.  Are you still seeing it with v3 of the patch?

Patch-set: 2