Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Also had a failure in the overnight run, which also was undertesting because I forgot to 'go clean -cache' between iterations.  Perhaps it was caused by that typo -- will try again with the fix, aimed at the failure case (in runtime): 
> 
> This is looking good to me.  I've been testing v2 of the patch locally with the typo fixed for five hours now.  The patch seems to be functioning correctly, in that instructions are being correctly padded, and I haven't seen any crashes, even when testing with GOGC=1 and using taskset 01.  Having said that, I've been unable to reproduce the runtime crash you reported on my machine, even with the version of the patch with the typo.  Are you still seeing it with v3 of the patch?

That weird crash is unrelated to your CL or these patches; I can reproduce it at tip, not 100% sure it is not just a test timing out because GOGC=1 on a single processor is not going to run that fast.

I.e., I think if we put the Prog NOP splicing into yours and re-enabled padding for RET, we're good.
There's some other discussion about whether we remove the option and just leave it on all the time, but I think that is a separate CL.

Patch-set: 3
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>