Update patch set 13
Patch Set 13: Run-TryBot+1

(12 comments)

> Patch Set 11: Code-Review+1
> 
> (12 comments)
> 
> This seems fiddly and a little bit madness-inducing.
> Would it be simpler to declare that the last page was not available for allocation?

I agree about the fiddliness and madness-inducing part. saying the last page is not available is effectively saying the last arena is not available. on 32-bit that's 4 MiB. the question is, is that enough to matter? we'd also need to map memory in a loop if we manage to get the top of the address space, though I guess that would be simpler than all this.

by the way, it should be at least a little bit less madness-inducing since you last looked at it. some of the fiddliness has gone away. perhaps addrRanges should be pointer+len instead of inclusive bounds? that might be simpler.

Patch-set: 13
Reviewer: Gerrit User 25391 <25391@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Run-TryBot=+1