runtime: avoid overflow with mheap.curArena
Currently mheap.curArena represents an address range using an inclusive
base address and exclusive top address, but the exclusive top address
could overflow if the current arena (or contiguous set of arenas) are
up against the top of the address space.

Instead, make curArena manage a base address and two offsets relative to
that base. This way, the overflow is eliminated, because the only way
those offsets could overflow is if the base is zero, but that isn't a
valid base.

For #35954.

Change-Id: I2d4ecbb1ccbd43da55979cc721f0cd8d1757add2