Update patch set 32
Patch Set 32:

(13 comments)

Getting there. Apologies for being pedantic, but this is a central package of the std library so we need to be extra diligent.

After this round we should be close or hopefully done. Thanks for doing this.

Patch-set: 32
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>