net/http: only support "chunked" in inbound Transfer-Encoding headers
This is a security hardening measure against HTTP request smuggling
prompted by an external reporter.

We weren't parsing things correctly anyway, allowing "identity" to be
combined with "chunked", and ignoring any Transfer-Encoding header past
the first. This is a delicate security surface that already broke
before, just be strict and don't add complexity to support cases not
observed in the wild (nginx removed "identity" support [1] and multiple
TE header support [2]) and removed by RFC 7230 (see page 81).

It'd probably be good to also drop support for anything else than
"chunked" in outbound TE headers, as "identity" is not a thing anymore,
and we are probably off-spec for anything else than "chunked", but it
should not be a security concern, so leaving it for now. See #38867.

[1]: https://hg.nginx.org/nginx/rev/fe5976aae0e3
[2]: https://hg.nginx.org/nginx/rev/aca005d232ff

DO NOT SUBMIT: waiting to hear from the reporter if and how they'd like
    to be acknowledged.

Change-Id: If17d0827f9c6167a0b19a158e2bc5844ec803288