cmd/go: disconnect GOAMD64 environment variable
This disconnects the GOAMD64 environment variable and removes
its documentation.  The logic remains against the possibility
that we need to restore the (option of) original alignment
behavior.

Note that function alignment is also dependent on the value of
the (disconnected) flag; it is 32 for aligned jumps, 16 if not.
When the flag-dependent logic is removed, it will be 32.

Change-Id: Ic41c0b9833d2e8a31fa3ce8067d92aa2f165bf72