cmd/go: default GOBIN to GOPATH[0]/bin for packages outside GOPATH
GOBIN now defaults to GOPATH[0] (if GOPATH is not empty).

'go env GOBIN' will print that default so this directory may easily be
used in scripting and setup instructions.

In GOPATH mode, when GOBIN is not set explicitly, packages within
a GOPATH root will still be installed in the root's bin directory.

This default will not be passed to subcommands. 'go tool dist'
(cmd/dist) gets confused, since it defaults GOBIN to $GOROOT/bin.

Fixes #23439

Change-Id: I647e20ee51d37334bb390855190648189874d7c1