Update patch set 3
Patch Set 3: Run-TryBot+1

This failed on the js-wasm builder because the default GOBIN was passed explicitly to 'go tool dist', but dist gets confused in the test environment when GOBIN is not $GOROOT/bin.

In general, I'm a little worried that we need several special cases for when GOBIN is not set explicitly. For installing binaries, it would be better to just use $GOPATH[0]/bin for files outside of GOPATH without changing GOBIN. But that would make 'go env GOBIN' less useful for scripting.

Patch-set: 3
Reviewer: Gerrit User 17092 <17092@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Run-TryBot=+1